# Conflicts:
#	SWP490_G9_PE/TnR_SS.Domain/Supervisor/TnR_SSSupervisorTimeKeeping.cs

+++ SWP490_G9_PE/TnR_SS.Domain/ApiModels/TimeKeepingModel/TimeKeepingApiModel.cs
@@ -12,7 +12,7 @@ namespace TnR_SS.Domain.ApiModels.TimeKeepingModel
     {
         public int ID { get; set; }
         public DateTime WorkDay { get; set; }
-        public int Status { get; set; }
+        public double Status { get; set; }
         public double Money { get; set; }
         public TimeKeepingNote Note { get; set; }
         public string EmpName { get; set; }
